requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
idna==3.4
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.0
packaging==24.2
pillow==11.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.28.1
six==1.17.0
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.13.0
urllib3==1.26.13
